Checking versions before push.
Capfire will now compare the currently deployed version with your local version, if thy are the same commit-message it will present you an error, and alert you of your boneheaded move. (It will only do this on a `before update_code` so deploy:migrate etc should be left alone.)

Also moved a lot of the logic from the capistrano script to the capfire-class. And moving from simple `puts` to capistrano's own logger class. All in all a maintenance update with some added features.